requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
cachetools==6.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.4.4
pillow==11.2.1
propcache==0.3.2
pycryptodomex==3.23.0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
steam==1.4.4
urllib3==2.4.0
vdf==3.4
yarl==1.20.1
